pyproject.toml
[tool.poetry]
name = "crypto-trader"
version = "0.1.0"
description = "Automatically trades cryptocurrency using set algorithms"
authors = ["Jason Park"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ccxt = "^4.3.58"
pandas = "^2.2.2"
websockets = "^12.0"
asyncio = "^3.4.3"
streamlit = "^1.36.0"
dash = "^2.17.1"
requests = "^2.32.3"
pyyaml = "^6.0.1"
jwt = "^1.3.1"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
pyorbital = "^1.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
